+++ src/rs_driver/driver/input/input_factory.hpp
@@ -31,16 +31,16 @@ THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 *********************************************************************************************************************/
 
 #pragma once
+
 #include <rs_driver/driver/input/input.hpp>
 #include <rs_driver/driver/input/input_raw.hpp>
+#include <rs_driver/driver/input/input_raw_jumbo.hpp>
+#include <rs_driver/driver/input/input_sock.hpp>
+#include <rs_driver/driver/input/input_sock_jumbo.hpp>
+
 #ifdef ENABLE_PCAP_PARSE
 #include <rs_driver/driver/input/input_pcap.hpp>
-#endif
-
-#ifdef __linux__
-#include <rs_driver/driver/input/unix/input_sock.hpp>
-#elif _WIN32
-#include <rs_driver/driver/input/win/input_sock.hpp>
+#include <rs_driver/driver/input/input_pcap_jumbo.hpp>
 #endif
 
 namespace robosense
@@ -51,11 +51,11 @@ namespace lidar
 class InputFactory
 {
 public:
-  static std::shared_ptr<Input> createInput(InputType type, const RSInputParam& param,
+  static std::shared_ptr<Input> createInput(InputType type, const RSInputParam& param, bool isJumbo,
       double sec_to_delay, std::function<void(const uint8_t*, size_t)>& cb_feed_pkt);
 };
 
-inline std::shared_ptr<Input> InputFactory::createInput(InputType type, const RSInputParam& param, 
+inline std::shared_ptr<Input> InputFactory::createInput(InputType type, const RSInputParam& param, bool isJumbo,
     double sec_to_delay, std::function<void(const uint8_t*, size_t)>& cb_feed_pkt)
 {
   std::shared_ptr<Input> input;
@@ -64,23 +64,36 @@ inline std::shared_ptr<Input> InputFactory::createInput(InputType type, const RS
   {
     case InputType::ONLINE_LIDAR:
       {
-        input = std::make_shared<InputSock>(param);
+        if (isJumbo)
+          input = std::make_shared<InputSockJumbo>(param);
+        else
+          input = std::make_shared<InputSock>(param);
       }
       break;
 
 #ifdef ENABLE_PCAP_PARSE
     case InputType::PCAP_FILE:
       {
-        input = std::make_shared<InputPcap>(param, sec_to_delay);
+        if (isJumbo)
+          input = std::make_shared<InputPcapJumbo>(param, sec_to_delay);
+        else
+          input = std::make_shared<InputPcap>(param, sec_to_delay);
       }
       break;
 #endif
 
     case InputType::RAW_PACKET:
       {
-        std::shared_ptr<InputRaw> inputRaw = std::make_shared<InputRaw>(param);
+        std::shared_ptr<InputRaw> inputRaw;
+
+        if (isJumbo)
+          inputRaw = std::make_shared<InputRawJumbo>(param);
+        else
+          inputRaw = std::make_shared<InputRaw>(param);
+
         cb_feed_pkt = std::bind(&InputRaw::feedPacket, inputRaw, 
             std::placeholders::_1, std::placeholders::_2);
+
         input = inputRaw;
       }
       break;